720- additional cucumbers steps for random

+++ generator/src/test/java/com/scottlogic/deg/generator/cucumber/steps/GeneralTestStep.java
@@ -222,7 +222,7 @@ public class GeneralTestStep {
     @Then("{long} rows of data are generated")
     public void theExpectedNumberOfRowsAreGenerated(long expectedNumberOfRows) {
         List <List<Object>> data = cucumberTestHelper.generateAndGetData();
-        Assert.assertThat("Unexpected number of rows returned", data.size(), equalTo(expectedNumberOfRows));
+        Assert.assertThat("Unexpected number of rows returned", (long)data.size(), equalTo(expectedNumberOfRows));
     }
 
     @Given("the generator can generate at most {long} rows")
